change the way in which demo DB is initialised: * instead of checking for 'demo_initialised', check if demo DB exists: if not, return error * don't log out the error if demo DB not initialised (this is part of the flow and not an error per se -- to prevent any, would be, crashlitics) * move db errors to 'cr-sqlite/errors.ts'
