+++ apps/web-client/src/lib/db/cr-sqlite/errors.ts
@@ -30,3 +30,44 @@ export class OutOfStockError extends Error {
 		this.invalidTransactions = invalidTransactions;
 	}
 }
+
+export class ErrDBCorrupted extends Error {
+	constructor(message: string) {
+		super(message);
+		this.name = "ErrDBCorrupted";
+	}
+}
+
+type ErrDBSchemaMismatchPayload = { wantName: string; wantVersion: bigint; gotName: string; gotVersion: bigint };
+export class ErrDBSchemaMismatch extends Error {
+	wantName: string;
+	wantVersion: bigint;
+
+	gotName: string;
+	gotVersion: bigint;
+
+	constructor({ wantName, wantVersion, gotName, gotVersion }: ErrDBSchemaMismatchPayload) {
+		const message = [
+			"DB name/schema mismatch:",
+			`  req name: ${wantName}, got name: ${gotName}`,
+			`  req version: ${wantVersion}, got version: ${gotVersion}`
+		].join("\n");
+
+		super(message);
+
+		this.name = "ErrDBSchemaMismatch";
+
+		this.wantName = wantName;
+		this.wantVersion = wantVersion;
+
+		this.gotName = gotName;
+		this.gotVersion = gotVersion;
+	}
+}
+
+export class ErrDemoDBNotInitialised extends Error {
+	constructor() {
+		super("Demo DB not initialised");
+		this.name = "ErrDemoDBNotInitialised";
+	}
+}
