+++ apps/web-client/src/routes/+layout.ts
@@ -16,16 +16,17 @@ import { setLocale } from "@librocco/shared/i18n-svelte";
 import { loadedLocales } from "@librocco/shared/i18n-util";
 import { locales } from "@librocco/shared/i18n-util";
 
-import { DEFAULT_LOCALE, DEFAULT_VFS, IS_DEMO, IS_E2E } from "$lib/constants";
+import { DEFAULT_LOCALE, DEFAULT_VFS, DEMO_DB_NAME, IS_DEMO, IS_E2E } from "$lib/constants";
 
 import { appPath } from "$lib/paths";
 import { newPluginsInterface } from "$lib/plugins";
 import { getDB } from "$lib/db/cr-sqlite";
-import { ErrDBCorrupted, ErrDBSchemaMismatch, ErrDemoDBNotInitialised } from "$lib/db/cr-sqlite/db";
+import { ErrDBCorrupted, ErrDBSchemaMismatch, ErrDemoDBNotInitialised } from "$lib/db/cr-sqlite/errors";
 import { validateVFS, type VFSWhitelist } from "$lib/db/cr-sqlite/core/vfs";
 
 import { updateTranslationOverrides } from "$lib/i18n-overrides";
 import { DEMO_VFS } from "$lib/db/cr-sqlite/core/constants";
+import { checkOPFSFileExists } from "$lib/db/cr-sqlite/core/utils";
 
 // Paths which are valid (shouldn't return 404, but don't have any content and should get redirected to the default route "/#/stock/")
 const redirectPaths = ["", "/", "/#", "/#/"].map((path) => `${base}${path}`);
@@ -74,15 +75,14 @@ export const load: LayoutLoad = async ({ url }) => {
 		try {
 			// DEMO section
 			if (IS_DEMO) {
+				// Check if the DB exists in OPFS
+				if (!(await checkOPFSFileExists(DEMO_DB_NAME))) {
+					throw new ErrDemoDBNotInitialised();
+				}
+
 				// In demo mode we use the hardcoded VFS
 				const vfs = DEMO_VFS;
 				const dbCtx = await getInitializedDB(get(dbid), vfs);
-
-				// Perform additional check - is demo DB empty?
-				const [res] = await dbCtx.db.execA("SELECT value FROM crsql_master WHERE key = 'demo_initialised'");
-				if (!res?.length || res[0] != "true") {
-					throw new ErrDemoDBNotInitialised();
-				}
 
 				return { dbCtx, plugins, error: null };
 			}
@@ -94,9 +94,13 @@ export const load: LayoutLoad = async ({ url }) => {
 			const dbCtx = await getInitializedDB(get(dbid), vfs);
 			return { dbCtx, plugins, error: null };
 		} catch (err) {
+			if (err instanceof ErrDemoDBNotInitialised) {
+				return { dbCtx: null, plugins, error: err as Error };
+			}
+
 			console.error("Error initializing DB", err);
 			// If know error, return it (it will ba handled in the shown dialog)
-			if (err instanceof ErrDBCorrupted || err instanceof ErrDBSchemaMismatch || err instanceof ErrDemoDBNotInitialised) {
+			if (err instanceof ErrDBCorrupted || err instanceof ErrDBSchemaMismatch) {
 				return { dbCtx: null, plugins, error: err as Error };
 			}
